Excepciones y respuestas mejoradas

+++ src/Base/DynamicModel.php
@@ -2,8 +2,11 @@
 
 namespace Jenyus\Base;
 
+use Exception;
+use InvalidArgumentException;
 use Jenyus\Base\Util\ModelFormat;
 use PDO;
+use PDOException;
 
 class DynamicModel
 {
@@ -17,9 +20,14 @@ class DynamicModel
      * Constructor de la clase DynamicModel.
      * 
      * @param PDO $conexion Objeto PDO para la conexión a la base de datos.
+     * @param $table Puedes definir opcionalmemte la propiedad table, al iniciar tu objeto.
+     * @throws PDOException Si no se proporciona una instancia válida de PDO.
      */
     public function __construct(PDO $conexion)
     {
+        if (!$conexion instanceof PDO) {
+            throw new InvalidArgumentException("Error in Jenyus\Base\DynamicModel: Se requiere una instancia válida de PDO.", 422);
+        }
         $this->conexion = $conexion;
     }
 
@@ -29,10 +37,24 @@ class DynamicModel
      * @param string $table Nombre de la tabla en la base de datos.
      * @return $this
      */
-    public function table($table)
+    public function SetTable($table)
     {
+        if (!is_string($table)) {
+            throw new InvalidArgumentException("Error in Jenyus\Base\DynamicModel: 'table' property must be a string. Please provide a valid table name.", 422);
+        }
         $this->table = $table;
         return $this;
+    }
+
+    /**
+     * @return $this->table value del objeto actual
+     */
+    public function getTable()
+    {
+        if (!$this->table || empty($this->table) || $this->table == '') {
+            throw new InvalidArgumentException("Error in Jenyus\Base\DynamicModel: Property 'table' cannot be null. Please call the 'table()' method with a valid table name before performing queries.", 422);
+        }
+        return $this->table;
     }
 
     /**
@@ -43,8 +65,12 @@ class DynamicModel
      */
     public function query($sql)
     {
-        $this->query = $this->conexion->query($sql);
-        return $this;
+        try {
+            $this->query = $this->conexion->query($sql);
+            return $this;
+        } catch (PDOException $e) {
+            throw new PDOException('Error in Jenyus\Base\DynamicModel: ' . $e->getMessage(), 500);
+        }
     }
 
     /**
@@ -54,7 +80,18 @@ class DynamicModel
      */
     public function first()
     {
-        return $this->query->fetch(\PDO::FETCH_ASSOC);
+        try {
+            if (!$this->query) {
+                throw new \PDOException("Error in Jenyus\Base\DynamicModel: No se ha ejecutado ninguna consulta previamente.");
+            }
+            $result = $this->query->fetch(\PDO::FETCH_ASSOC);
+            if (!$result) {
+                return false;
+            }
+            return $result;
+        } catch (PDOException $e) {
+            throw new PDOException('Error in Jenyus\Base\DynamicModel: ' . $e->getMessage());
+        }
     }
 
     /**
@@ -64,7 +101,18 @@ class DynamicModel
      */
     public function get()
     {
-        return $this->query->fetchAll(\PDO::FETCH_ASSOC);
+        try {
+            if (!$this->query) {
+                throw new \PDOException("No se ha ejecutado ninguna consulta previamente.");
+            }
+            $result = $this->query->fetchAll(\PDO::FETCH_ASSOC);
+            if (!$result) {
+                return false;
+            }
+            return $result;
+        } catch (PDOException $e) {
+            throw new PDOException('Error in Jenyus\Base\DynamicModel: ' . $e->getMessage());
+        }
     }
 
     /**
@@ -75,6 +123,12 @@ class DynamicModel
      */
     public function all($columns = ['*'])
     {
+
+        if (!is_array($columns)) {
+            throw new InvalidArgumentException("Error in Jenyus\Base\DynamicModel: The argument must be array", 422);
+        }
+
+        $this->getTable();
         // Convertir el array de columnas en una cadena separada por comas
         $columnsStr = implode(', ', $columns);
         // Construir la consulta preparada
@@ -84,8 +138,8 @@ class DynamicModel
             $stmt->execute();
             $this->query = $stmt;
             return $this;
-        } catch (\PDOException $e) {
-            return [false, $e->getMessage()];
+        } catch (PDOException $e) {
+            throw new PDOException('Error in Jenyus\Base\DynamicModel: ' . $e->getMessage(), 500);
         }
     }
 
@@ -100,6 +154,11 @@ class DynamicModel
      */
     public function where($column, $value, $operator = "=", $columns = ['*'])
     {
+        if (!is_array($columns)) {
+            throw new InvalidArgumentException("Error in Jenyus\Base\DynamicModel: The argument must be array", 422);
+        }
+
+        $this->getTable();
         // Convertir el array de columnas en una cadena separada por comas
         $columnsStr = implode(', ', $columns);
         // Construir la consulta preparada
@@ -110,9 +169,11 @@ class DynamicModel
             $stmt->bindParam(':value', $value, $param_type);
             $stmt->execute();
             $this->query = $stmt;
-            return [true, $this];
-        } catch (\PDOException $e) {
-            return [false, $e->getMessage()];
+            return $this;
+        } catch (InvalidArgumentException $e) {
+            throw new InvalidArgumentException("Error in Jenyus\Base\DynamicModel: The argument is not correct", 422);
+        } catch (PDOException $e) {
+            throw new PDOException("Error in Jenyus\Base\DynamicModel: " . $e->getMessage());
         }
     }
 
@@ -127,19 +188,25 @@ class DynamicModel
      */
     public function find($value, $columns = ['*'], $operator = '=', $column = 'id')
     {
-        // Convertir el array de columnas en una cadena separada por comas
-        $columnsStr = implode(', ', $columns);
-        // Construir la consulta preparada
-        $sql = "SELECT {$columnsStr} FROM {$this->table} WHERE {$column} {$operator} :value";
+        if (!is_array($columns)) {
+            throw new InvalidArgumentException("The argument must be array", 422);
+        }
+
+        $this->getTable();
+
         try {
+            // Convertir el array de columnas en una cadena separada por comas
+            $columnsStr = implode(', ', $columns);
+            // Construir la consulta preparada
+            $sql = "SELECT {$columnsStr} FROM {$this->table} WHERE {$column} {$operator} :value";
             $stmt = $this->conexion->prepare($sql);
             $param_type = is_int($value) ? \PDO::PARAM_INT : \PDO::PARAM_STR;
             $stmt->bindParam(':value', $value, $param_type);
             $stmt->execute();
             $this->query = $stmt;
-            return [true, $this];
-        } catch (\PDOException $e) {
-            return [false, $e->getMessage()];
+            return $this->first();
+        } catch (PDOException $e) {
+            throw new PDOException("Error en el servidor: " . $e->getMessage(), 500);
         }
     }
 
@@ -147,30 +214,35 @@ class DynamicModel
      * Inserta nuevos registros en la tabla especificada.
      * 
      * @param array $columns array asociativo con las columnas y valores
-     * @return array Arreglo con el resultado de la inserción.
+     * @return array id del registro insertado.
      */
     public function insert($columns = [], $dateTime = true)
     {
-        // Añadir 'created_at' a las columnas y su valor correspondiente a los valores
-        if($dateTime){
-            $columns['created_at'] = $this->basicCurrentFormatDate();
-        }
-        $columnNames = '';
-        $placeholders = '';
-        $values = [];
-
-        foreach ($columns as $key => $value) {
-            $columnNames .= $key . ', ';
-            $placeholders .= ':' . $key . ', ';
-            $values[':' . $key] = $value;
+        $this->getTable();
+        if (!is_array($columns)) {
+            throw new InvalidArgumentException("Error in Jenyus\Base\DynamicModel: The argument must be associative array", 422);
         }
 
-        // Eliminar las comas y espacios al final de las cadenas
-        $columnNames = rtrim($columnNames, ', ');
-        $placeholders = rtrim($placeholders, ', ');
-
-        $sql = "INSERT INTO {$this->table} ({$columnNames}) VALUES ({$placeholders})";
         try {
+            // Añadir 'created_at' a las columnas y su valor correspondiente a los valores
+            if ($dateTime) {
+                $columns['created_at'] = $this->basicCurrentFormatDate();
+            }
+            $columnNames = '';
+            $placeholders = '';
+            $values = [];
+
+            foreach ($columns as $key => $value) {
+                $columnNames .= $key . ', ';
+                $placeholders .= ':' . $key . ', ';
+                $values[':' . $key] = $value;
+            }
+
+            // Eliminar las comas y espacios al final de las cadenas
+            $columnNames = rtrim($columnNames, ', ');
+            $placeholders = rtrim($placeholders, ', ');
+
+            $sql = "INSERT INTO {$this->table} ({$columnNames}) VALUES ({$placeholders})";
             $stmt = $this->conexion->prepare($sql);
             // Asignar valores a los placeholders
             foreach ($values as $key => $value) {
@@ -178,15 +250,17 @@ class DynamicModel
                 $stmt->bindValue($key, $value, $param_type);
             }
             $stmt->execute();
+
             if ($stmt->rowCount() > 0) {
-                $id = $this->conexion->lastInsertId();
-                return [true, $id];
+                return $this->conexion->lastInsertId();
             }
-            // Si no se insertó ninguna fila (rowCount <= 0), puede manejarlo según tus requerimientos.
-            return [false, 'No se insertaron filas'];
-        } catch (\PDOException $e) {
-            // Manejar errores de PDO
-            return [false, $e->getMessage()];
+
+            throw new \RuntimeException('Jenyus\Base\DynamicModel: No se insertaron filas', 204);
+
+        } catch (PDOException $e) {
+            throw new PDOException("Error in Jenyus\Base\DynamicModel: " . $e->getMessage(), 500);
+        } catch (\Exception $e) {
+            throw new Exception("Error in Jenyus\Base\DynamicModel: " . $e->getMessage(), 500);
         }
     }
 
@@ -202,28 +276,46 @@ class DynamicModel
      */
     public function update($columns = [], $value, $operator = '=', $column = 'id')
     {
-        // Construir la cadena SET
-        $sets = "";
-        foreach ($columns as $key => $c) {
-            $sets .= "$key = :$key, ";
+        $data = $this->find($value, ['id'], $operator, $column);
+        if (!$data) {
+            throw new PDOException("Error in Jenyus\Base\DynamicModel: el recurso no existe", 404);
         }
-        $sets = rtrim($sets, ', ');
-        $sql = "UPDATE {$this->table} SET {$sets} WHERE $column $operator :whereValue";
+        $this->getTable();
+        if (!is_array($columns)) {
+            throw new InvalidArgumentException("Error in Jenyus\Base\DynamicModel: The argument must be associative array");
+        }
 
         try {
+            // Construir la cadena SET
+            $sets = "";
+            foreach ($columns as $key => $c) {
+                $sets .= "$key = :$key, ";
+            }
+            $sets = rtrim($sets, ', ');
+
+            $sql = "UPDATE {$this->table} SET {$sets} WHERE $column $operator :whereValue";
             $stmt = $this->conexion->prepare($sql);
+
             // Enlazar los valores de las columnas
             foreach ($columns as $key => $val) {
                 $param_type = is_int($val) ? \PDO::PARAM_INT : \PDO::PARAM_STR;
                 $stmt->bindValue(":$key", $val, $param_type);
             }
+
             // Enlazar el valor para la cláusula WHERE
             $param_type = is_int($value) ? \PDO::PARAM_INT : \PDO::PARAM_STR;
             $stmt->bindValue(":whereValue", $value, $param_type);
             $stmt->execute();
-            return [$stmt->rowCount() > 0, 'Actualizado'];
-        } catch (\PDOException $e) {
-            return [false, $e->getMessage()];
+
+            if ($stmt->rowCount() > 0) {
+                return $this->find($value, ['id']);
+            }
+
+            return false;
+        } catch (PDOException $e) {
+            throw new PDOException("Error in Jenyus\Base\DynamicModel: " . $e->getMessage(), 500);
+        } catch (\Exception $e) {
+            throw new Exception("Error interno: " . $e->getMessage(), 500);
         }
     }
 
@@ -237,15 +329,30 @@ class DynamicModel
      */
     public function delete($value, $operator = '=', $column = 'id')
     {
-        $sql = "DELETE FROM {$this->table} WHERE {$column} {$operator} :value";
+        $this->getTable();
+
+        $user = $this->find($value, ['id'], $operator, $column);
+
+        if (!$user) {
+            throw new PDOException("Error in Jenyus\Base\DynamicModel: el registro no existe", 404);
+            return;
+        }
+
         try {
+            $sql = "DELETE FROM {$this->table} WHERE {$column} {$operator} :value";
             $stmt = $this->conexion->prepare($sql);
             $param_type = is_int($value) ? \PDO::PARAM_INT : \PDO::PARAM_STR;
             $stmt->bindParam(':value', $value, $param_type);
             $stmt->execute();
-            return [$stmt->rowCount() > 0, 'Eliminado'];
-        } catch (\PDOException $e) {
-            return [false, $e->getMessage()];
+            if ($stmt->rowCount() > 0) {
+                return $this->conexion->lastInsertId();
+            }
+
+            throw new \RuntimeException('No se actualizaron filas');
+        } catch (PDOException $e) {
+            throw new PDOException("Error in Jenyus\Base\DynamicModel: " . $e->getMessage(), 500);
+        } catch (\Exception $e) {
+            throw new Exception("Error in Jenyus\Base\DynamicModel: " . $e->getMessage(), 500);
         }
     }
 }
